pyproject.toml
[tool.poetry]
name = "drf-api-permissions-postgres"
version = "0.1.0"
description = ""
authors = ["Rawan199812 <rawan.13579@icloud.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.4"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]
black = "^21.6b0"
flake8 = "^3.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
